Move maximum statistics population to `LegacyScoreDecoder`
