+++ osu.Game/Scoring/ScoreImporter.cs
@@ -17,7 +17,6 @@ using osu.Game.Scoring.Legacy;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests;
 using osu.Game.Online.API.Requests.Responses;
-using osu.Game.Rulesets.Scoring;
 using Realms;
 
 namespace osu.Game.Scoring
@@ -91,8 +90,6 @@ namespace osu.Game.Scoring
             ArgumentNullException.ThrowIfNull(model.BeatmapInfo);
             ArgumentNullException.ThrowIfNull(model.Ruleset);
 
-            PopulateMaximumStatistics(model);
-
             if (string.IsNullOrEmpty(model.StatisticsJson))
                 model.StatisticsJson = JsonConvert.SerializeObject(model.Statistics);
 
@@ -108,70 +105,6 @@ namespace osu.Game.Scoring
                 model.LegacyTotalScore = model.TotalScore;
                 StandardisedScoreMigrationTools.UpdateFromLegacy(model, beatmaps());
             }
-        }
-
-        /// <summary>
-        /// Populates the <see cref="ScoreInfo.MaximumStatistics"/> for a given <see cref="ScoreInfo"/>.
-        /// </summary>
-        /// <param name="score">The score to populate the statistics of.</param>
-        public void PopulateMaximumStatistics(ScoreInfo score)
-        {
-            Debug.Assert(score.BeatmapInfo != null);
-
-            if (score.MaximumStatistics.Select(kvp => kvp.Value).Sum() > 0)
-                return;
-
-            var beatmap = score.BeatmapInfo!.Detach();
-            var ruleset = score.Ruleset.Detach();
-            var rulesetInstance = ruleset.CreateInstance();
-            var scoreProcessor = rulesetInstance.CreateScoreProcessor();
-
-            Debug.Assert(rulesetInstance != null);
-
-            // Populate the maximum statistics.
-            HitResult maxBasicResult = rulesetInstance.GetHitResults()
-                                                      .Select(h => h.result)
-                                                      .Where(h => h.IsBasic()).MaxBy(scoreProcessor.GetBaseScoreForResult);
-
-            foreach ((HitResult result, int count) in score.Statistics)
-            {
-                switch (result)
-                {
-                    case HitResult.LargeTickHit:
-                    case HitResult.LargeTickMiss:
-                        score.MaximumStatistics[HitResult.LargeTickHit] = score.MaximumStatistics.GetValueOrDefault(HitResult.LargeTickHit) + count;
-                        break;
-
-                    case HitResult.SmallTickHit:
-                    case HitResult.SmallTickMiss:
-                        score.MaximumStatistics[HitResult.SmallTickHit] = score.MaximumStatistics.GetValueOrDefault(HitResult.SmallTickHit) + count;
-                        break;
-
-                    case HitResult.IgnoreHit:
-                    case HitResult.IgnoreMiss:
-                    case HitResult.SmallBonus:
-                    case HitResult.LargeBonus:
-                        break;
-
-                    default:
-                        score.MaximumStatistics[maxBasicResult] = score.MaximumStatistics.GetValueOrDefault(maxBasicResult) + count;
-                        break;
-                }
-            }
-
-            if (!score.IsLegacyScore)
-                return;
-
-#pragma warning disable CS0618
-            // In osu! and osu!mania, some judgements affect combo but aren't stored to scores.
-            // A special hit result is used to pad out the combo value to match, based on the max combo from the difficulty attributes.
-            var calculator = rulesetInstance.CreateDifficultyCalculator(beatmaps().GetWorkingBeatmap(beatmap));
-            var attributes = calculator.Calculate(score.Mods);
-
-            int maxComboFromStatistics = score.MaximumStatistics.Where(kvp => kvp.Key.AffectsCombo()).Select(kvp => kvp.Value).DefaultIfEmpty(0).Sum();
-            if (attributes.MaxCombo > maxComboFromStatistics)
-                score.MaximumStatistics[HitResult.LegacyComboIncrease] = attributes.MaxCombo - maxComboFromStatistics;
-#pragma warning restore CS0618
         }
 
         // Very naive local caching to improve performance of large score imports (where the username is usually the same for most or all scores).
